Super.DisplayException for drawn
