+++ src/samples/Sandbox/View1.xaml.cs
@@ -1,0 +1,17 @@
+﻿namespace Sandbox
+{
+    public partial class View1 : SkiaLayout
+    {
+        public View1()
+        {
+            try
+            {
+                InitializeComponent();
+            }
+            catch (Exception e)
+            {
+                Super.DisplayException(this, e);
+            }
+        }
+    }
+}
